pyproject.toml
[project]
name = "psoriasis-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.120.3",
    "ipywidgets>=8.1.8",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "mlflow>=3.5.1",
    "nest-asyncio>=1.6.0",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "sqlalchemy>=2.0.44",
    "wordcloud>=1.9.4",
]
authors = [
    { name = "Umair Ashraf", email = "umairashrafbsse@gmail.com" },
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
